Move out of index.
