+++ src/pages/index.tsx
@@ -1,38 +1,9 @@
 import Head from "next/head";
-import { Backdrop, CircularProgress, Typography } from "@mui/material";
-import { useState } from "react";
-import { useMutation, useQuery } from "@apollo/client";
 
 import styles from "@/styles/Home.module.css";
-import { queries } from "@kenk2/graphql";
-import { TodoCard, TodoInput } from "@kenk2/components";
-import Todo from "@kenk2/types";
+import TodoContainer from "../components/TodoContainer";
 
 export default function Home() {
-  const { loading, data } = useQuery(queries.GET_TODOS);
-  const [deleteId, setDeleteId] = useState<undefined | number>();
-  const [editId, setEditId] = useState<undefined | number>();
-  const [deleteTodos] = useMutation(queries.DELETE_TODO, {
-    onCompleted: () => setDeleteId(undefined),
-    refetchQueries: [queries.GET_TODOS],
-  });
-  const [editTodos] = useMutation(queries.UPDATE_TODO, {
-    onCompleted: () => setEditId(undefined),
-    refetchQueries: [queries.GET_TODOS],
-  });
-
-  const handleDelete = (id: number) => {
-    setDeleteId(id);
-    if (editId === id) {
-      setEditId(undefined);
-    }
-    deleteTodos({ variables: { id } });
-  };
-
-  const handleEdit = (id: number, text: string) => {
-    editTodos({ variables: { id, text } });
-  };
-
   return (
     <>
       <Head>
@@ -42,22 +13,7 @@ export default function Home() {
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <main className={styles.main}>
-        <Typography sx={{ marginBottom: "8px" }}>GraphQL Todos</Typography>
-        <Backdrop open={loading}>
-          <CircularProgress />
-        </Backdrop>
-        <TodoInput />
-        {data?.todos?.map((todo: Todo) => (
-          <TodoCard
-            todo={todo}
-            key={todo.id}
-            onEditToggle={(id) => setEditId(id)}
-            onDelete={handleDelete}
-            onEditSave={handleEdit}
-            editting={todo.id === editId}
-            deleting={todo.id === deleteId}
-          />
-        ))}
+        <TodoContainer />
       </main>
     </>
   );
